+++ app/src/main/res/layout/activity_sips_gas.xml
@@ -1,0 +1,121 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".sips.ActivitySipsGas">
+
+    <ScrollView
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:background="@drawable/pngwing_com">
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:layout_gravity="top"
+            android:layout_margin="20dp"
+            android:gravity="center"
+            android:orientation="vertical"
+            tools:layout_editor_absoluteX="40dp"
+            tools:layout_editor_absoluteY="16dp">
+
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="196dp"
+                android:orientation="vertical">
+
+                <ImageView
+                    android:id="@+id/imageView3Gas"
+                    android:layout_width="182dp"
+                    android:layout_height="111dp"
+                    android:layout_gravity="center"
+                    app:srcCompat="@drawable/inergas_logo_transparente" />
+            </LinearLayout>
+
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="339dp"
+                android:layout_gravity="center"
+                android:orientation="vertical">
+
+                <EditText
+                    android:id="@+id/txtBuscaCupsGas"
+                    android:layout_width="wrap_content"
+                    android:layout_height="64dp"
+                    android:layout_margin="10dp"
+                    android:background="@drawable/th"
+                    android:ems="10"
+                    android:gravity="center"
+                    android:hint="@string/hintBuscaCupsGas"
+                    android:inputType="textPersonName"
+                    android:text="ES0224050004873564DG"
+                    android:textColor="@color/white"
+                    android:textColorHint="@color/white" />
+
+                <androidx.appcompat.widget.AppCompatButton
+                    android:id="@+id/btnBuscaCupsGas"
+                    android:layout_width="wrap_content"
+                    android:layout_height="72dp"
+                    android:layout_margin="10dp"
+                    android:layout_marginTop="20dp"
+                    android:background="@drawable/boton_verde_presionado"
+                    android:gravity="center"
+                    android:text="@string/btnBuscaCupsGas"
+                    android:textColor="@color/white" />
+
+                <TextView
+                    android:id="@+id/lblConsumoGas"
+                    android:layout_width="match_parent"
+                    android:layout_height="49dp"
+                    android:layout_margin="10dp"
+                    android:background="@drawable/fondogas"
+                    android:gravity="center"
+                    android:text="@string/lblConsumoCupsGas"
+                    android:textColor="@color/white"
+                    android:textSize="20sp" />
+
+                <EditText
+                    android:id="@+id/txtCupsConsumoGas"
+                    android:layout_width="match_parent"
+                    android:layout_height="64dp"
+                    android:layout_margin="10dp"
+                    android:background="@drawable/th"
+                    android:ems="10"
+                    android:gravity="center"
+                    android:inputType="numberDecimal"
+                    android:textColor="@color/white" />
+            </LinearLayout>
+
+            <LinearLayout
+                android:layout_width="371dp"
+                android:layout_height="match_parent"
+                android:layout_margin="10dp"
+                android:gravity="center"
+                android:orientation="horizontal">
+
+                <androidx.appcompat.widget.AppCompatButton
+                    android:id="@+id/btnAtrasBuscaCupsGas"
+                    android:layout_width="132dp"
+                    android:layout_height="55dp"
+                    android:layout_margin="10dp"
+                    android:background="@drawable/boton_verde"
+                    android:gravity="center"
+                    android:text="@string/btnAnteriorBuscaCupsGas"
+                    android:textColor="@color/white" />
+
+                <androidx.appcompat.widget.AppCompatButton
+                    android:id="@+id/btnSiguienteBuscaCupsGas"
+                    android:layout_width="129dp"
+                    android:layout_height="53dp"
+                    android:layout_margin="10dp"
+                    android:background="@drawable/boton_verde"
+                    android:gravity="center"
+                    android:text="@string/btnSiguienteBuscaCupsGas"
+                    android:textColor="@color/white" />
+            </LinearLayout>
+
+        </LinearLayout>
+    </ScrollView>
+</androidx.constraintlayout.widget.ConstraintLayout>
